Reorganise file for better cohesion.
