+++ tests/repositories/test_personal_info_repository.py
@@ -3,7 +3,8 @@ from unittest.mock import patch
 import pytest
 from sqlalchemy.exc import NoResultFound, SQLAlchemyError
 
-from app.repositories.personal_info_repository import get_person_from_db
+from app.repositories.applicant.personal_info_repository import \
+    get_person_from_db
 from tests.utilities.test_utilities import remove_test_user_from_db, \
     setup_test_user_in_db
 
@@ -29,7 +30,7 @@ def test_get_person_from_db_no_result(app_with_client):
 def test_get_person_from_db_sqlalchemy_error(app_with_client):
     app, _ = app_with_client
     with app.app_context():
-        with patch('app.models.person.Person.query') as mock_query:
+        with patch('app.models.applicant.person.Person.query') as mock_query:
             mock_query.filter_by.side_effect = SQLAlchemyError(
                     "A database error occurred")
             with pytest.raises(SQLAlchemyError) as e:
